+++ stanza/iq.go
@@ -108,6 +108,15 @@ func (iq IQ) Result(payload xml.TokenReader) xml.TokenReader {
 	return iq.Wrap(payload)
 }
 
+// Error returns a token reader that wraps the first element from payload in an
+// IQ stanza with the to and from attributes switched and the type set to
+// ErrorIQ.
+func (iq IQ) Error(err Error) xml.TokenReader {
+	iq.Type = ErrorIQ
+	iq.From, iq.To = iq.To, iq.From
+	return iq.Wrap(err.TokenReader())
+}
+
 // IQType is the type of an IQ stanza.
 // It should normally be one of the constants defined in this package.
 type IQType string